+++ src/gui/graphtitlebar.cpp
@@ -18,6 +18,14 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <QFileDialog>
+#include <QMessageBox>
+#include <QProgressDialog>
+
+#include "core/nodes_p.h"
+#include "core/project.h"
+#include "core/trial.h"
+
 #include "graphtitlebar.h"
 #include "ui_graphtitlebar.h"
 
@@ -27,12 +35,20 @@ GraphTitleBar::GraphTitleBar(const Experiment* exp, QDockWidget* parent)
     : BaseTitleBar(parent),
       m_ui(new Ui_GraphTitleBar),
       m_exp(exp),
+      m_bExportNodes(new QtMaterialIconButton(QIcon(":/icons/material/table_white_18"), this)),
       m_bSettings(new QtMaterialIconButton(QIcon(":/icons/material/settings_white_18"), this))
 {
     m_ui->setupUi(this);
 
-    m_bSettings->setIconSize(QSize(18,18));
-    m_bSettings->setColor(Qt::white);
+    m_bExportNodes->setToolTip("export nodes to file");
+    m_bExportNodes->setIconSize(QSize(22,18));
+    m_bExportNodes->setColor(m_iconColor);
+    m_ui->btns->addWidget(m_bExportNodes);
+    connect(m_bExportNodes, SIGNAL(clicked(bool)), SLOT(slotExporNodes()));
+
+    m_bSettings->setToolTip("graph settings");
+    m_bSettings->setIconSize(QSize(22,18));
+    m_bSettings->setColor(m_iconColor);
     m_ui->btns->addWidget(m_bSettings);
     connect(m_bSettings, SIGNAL(clicked(bool)), SIGNAL(openSettingsDlg()));
 
@@ -51,6 +67,53 @@ GraphTitleBar::GraphTitleBar(const Experiment* exp, QDockWidget* parent)
 GraphTitleBar::~GraphTitleBar()
 {
     delete m_ui;
+}
+
+void GraphTitleBar::slotExporNodes()
+{
+    if (m_exp->expStatus() == Status::Disabled ||
+            m_exp->expStatus() == Status::Invalid) {
+        QMessageBox::warning(this, "Exporting nodes",
+                "This experiment is invalid or has not been initialized yet.\n"
+                "Please, initialize the experiment and try again.");
+        return;
+    } else if (m_exp->expStatus() == Status::Running ||
+                    m_exp->expStatus() == Status::Queued) {
+        QMessageBox::warning(this, "Exporting nodes",
+                "Please, pause the experiment and try again.");
+        return;
+    }
+
+    const quint16 currTrial = m_ui->cbTrial->currentText().toUShort();
+    auto trial = m_exp->trial(currTrial);
+    if (!trial || trial->graph()->nodes().empty()) {
+        QMessageBox::warning(this, "Exporting nodes",
+                "Could not export the set of nodes.\n"
+                "Please, make sure this experiment is valid and that "
+                "there are nodes to be exported!");
+        return;
+    }
+
+    QString path = m_exp->project()->filepath();
+    path = QFileDialog::getSaveFileName(this, "Export Nodes",
+            path.isEmpty() ? QDir::homePath() : path, "Text Files (*.csv)");
+    if (path.isEmpty()) {
+        return;
+    }
+
+    QProgressDialog progressDlg("Exporting nodes", QString(), 0, trial->graph()->numNodes(), this);
+    progressDlg.setWindowModality(Qt::WindowModal);
+    progressDlg.setValue(0);
+    std::function<void(int)> progress = [&progressDlg](int p) { progressDlg.setValue(p); };
+
+    if (NodesPrivate::saveToFile(trial->graph()->nodes(), path, progress)) {
+        QMessageBox::information(this, "Exporting nodes",
+                "The set of nodes was saved successfully!\n" + path);
+    } else {
+        QMessageBox::warning(this, "Exporting nodes",
+                "ERROR! Unable to save the set of nodes at:\n"
+                + path + "\nPlease, make sure this directory is writable.");
+    }
 }
 
 void GraphTitleBar::slotRestarted()